+++ libs/material/popover-edit/src/index.ts
@@ -1,0 +1,20 @@
+export * from './lib/cdk/edit-event-dispatcher';
+export * from './lib/cdk/edit-ref';
+export * from './lib/cdk/focus-dispatcher';
+export * from './lib/cdk/form-value-container';
+export * from './lib/cdk/lens-directives';
+export * from './lib/cdk/popover-edit-module';
+export * from './lib/cdk/popover-edit-position-strategy-factory';
+export * from './lib/cdk/table-directives';
+
+// Private to Angular Components
+export { CELL_SELECTOR as _CELL_SELECTOR } from './lib/cdk/constants';
+export {
+  closest as _closest,
+  matches as _matches
+} from './lib/cdk/polyfill';
+
+export * from './lib/lens-directives';
+export * from './lib/popover-edit-module';
+export * from './lib/table-directives';
+export * from './lib/popover-edit-form.directive';