+++ libs/xml-parser/CHANGELOG.md
@@ -22,6 +22,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+# [12.1.0-alpha.0](https://gitlab.com/rxap/packages/compare/@rxap/xml-parser@12.1.0-next.1...@rxap/xml-parser@12.1.0-alpha.0) (2021-07-20)
+
+**Note:** Version bump only for package @rxap/xml-parser
+
+
+
+
+
 # [12.1.0-next.1](https://gitlab.com/rxap/packages/compare/@rxap/xml-parser@12.1.0-next.0...@rxap/xml-parser@12.1.0-next.1) (2021-07-20)